% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downlaod_inat_images.R
\name{downlaod_inat_images}
\alias{downlaod_inat_images}
\title{Download iNaturalist images}
\usage{
downlaod_inat_images(
  dat,
  size = "medium",
  outpath = ".",
  img_format = NULL,
  firstonly = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{dat}{data frame containing iNaturalist records}

\item{size}{image size. Values accepted are "small", "medium", "large", "original".
Deafault(medium).}

\item{outpath}{path to the output folder}

\item{img_format}{image formating parameters in the form of list. Detailed
documentation is in the function \code{\link{format_image}}}

\item{firstonly}{All images need to be downloaded or only first for each
iNaturalist record}

\item{verbose}{should messges on progress to be displayed}
}
\description{
Download iNaturalist images
}
\examples{
\dontrun{
myfmt <- NULL
myfmt[[1]] <- c("border","white","20x50")
myfmt[[2]] <- c("border","grey", "2x2")
myfmt[[3]] <- c("annotate",c('"\u00A9"', 'photoby','"   "'),"black", 20,"southeast")
myfmt[[4]] <- c("annotate",c('" "','scname','"\\\\n "', 'cname'),"black", 20,"northwest")
myfmt[[5]] <- c("annotate",c('" "','place','"\\\\n "', 'photodate'),"black", 20,"southwest")
myfmt[[6]] <- c("annotate",c('"iNat Id \\\\n "','obs_id','" "'),"black", 20,"northeast")

# Get some iNaturalist data for a butterfly Idea malabarica
inatdata <- get_inat_obs(query = "Idea malabarica",maxresults = 5)
downlaod_inat_images(inatdata,size="medium", outpath=".\\\\test\\\\", img_format=myfmt,
                     firstonly=TRUE,verbose=TRUE)
# Get some iNaturalist data for a user
inatdata <- get_inat_obs_user("vijaybarve", maxresults = 15)
downlaod_inat_images(inatdata,size="medium", outpath=".\\\\test\\\\", img_format=myfmt,
                     firstonly=TRUE,verbose=TRUE)
}
}
\seealso{
Other image functions: 
\code{\link{format_image}()}
}
\concept{image functions}
