% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_image.R
\name{format_image}
\alias{format_image}
\title{Download formated iNaturalist images}
\usage{
format_image(imginfo, img_no = 1, size = "medium", img_format = NULL)
}
\arguments{
\item{imginfo}{iNaturalist id of the record}

\item{img_no}{image number for the image for the iNaturalist record}

\item{size}{mage size. Values accepted are "small", "medium", "large", "original". Deafault(medium).}

\item{img_format}{image formating parameters in the form of list. Each item of the list is a
component to be added on the image. currently Border and Annotations are available. For Border
the componants of the list are \itemize{ \item{keyword -} {"border"} \item{color -} {color name}
\item{border dimentions -} {horizontal x verticle border in pixels.}}
For Annotations the componants of the list are \itemize{ \item{keyword -} {"annonate"}
\item{text -} { see bellow list of text components}
\item{color -} {color name}
\item{font size -} {Font size in pixel}
\item{placement -} {"southeast", "southwest", "northeast", "northwest"}}
Text annotation componants available from the iNaturalist records are as follows
\itemize{ \item{obs_id : Observation id by iNaturalist}
\item{photoby : Name of photographer}
\item{scname : Scientific name}
\item{canem : Common name}
\item{photodate : Date of record}
\item{place : Location }
}}
}
\description{
Download iNaturalist images for a recod and format them with metadata
}
\examples{
\dontrun{
myfmt <- NULL
myfmt[[1]] <- c("border","white","20x50")
myfmt[[2]] <- c("border","grey", "2x2")
myfmt[[3]] <- c("annotate",c('"\u00A9"', 'photoby','"   "'),"black", 20,"southeast")
myfmt[[4]] <- c("annotate",c('" "','scname','"\\\\n "', 'cname'),"black", 20,"northwest")
myfmt[[5]] <- c("annotate",c('" "','place','"\\\\n "', 'photodate'),"black", 20,"southwest")
myfmt[[6]] <- c("annotate",c('"iNat Id \\\\n "','obs_id','" "'),"black", 20,"northeast")
format_image(imginfo,img_no,size="medium",img_format=myfmt)
}
}
\seealso{
Other image functions: 
\code{\link{downlaod_inat_images}()}
}
\concept{image functions}
